Fix char array subscript warning

+++ librz/util/print.c
@@ -1790,7 +1790,7 @@ static bool issymbol(char c) {
 static bool check_arg_name(RzPrint *print, char *p, ut64 func_addr) {
 	if (func_addr && print->exists_var) {
 		int z;
-		for (z = 0; p[z] && (isalpha(p[z]) || isdigit(p[z]) || p[z] == '_'); z++) {
+		for (z = 0; p[z] && (isalpha((int)p[z]) || isdigit((int)p[z]) || p[z] == '_'); z++) {
 			;
 		}
 		char tmp = p[z];
